+++ src/lib/types/LabelValueOption.ts
@@ -1,7 +1,13 @@
+interface LabelValueOptionGroup {
+  name: string;
+  disabled?: boolean;
+}
+
 interface LabelValueOption {
   value?: string | number;
   label: string;
   disabled?: boolean;
+  group?: LabelValueOptionGroup;
 }
 
-export { LabelValueOption };
+export { LabelValueOption, LabelValueOptionGroup };