+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.12.0] - 2024-10-07
+
+### Added
+
+- The property `group` in the `LabelValueOption`
+- The possibility to group results both in the `AsyncTypeaheadInput` and `StaticTypeaheadInput` using the property `useGroupBy`
+
 ## [2.11.0] - 2024-10-04
 
 ### Added
@@ -409,7 +416,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Created package :tada:
 
-[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.11.0...HEAD
+[unreleased]: https://github.com/neolution-ch/react-hook-form-components/compare/2.12.0...HEAD
 [0.1.2]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/neolution-ch/react-hook-form-components/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.1.0
@@ -434,6 +441,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.4.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/neolution-ch/react-hook-form-components/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/neolution-ch/react-hook-form-components/releases/tag/0.2.0
+[2.12.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.11.0...2.12.0
 [2.11.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.10.0...2.11.0
 [2.10.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.9.0...2.10.0
 [2.9.0]: https://github.com/neolution-ch/react-hook-form-components/compare/2.8.1...2.9.0